+++ cspdk/si220/cells/__init__.py
@@ -1,4 +1,0 @@
-"""Cells."""
-
-from .containers import *
-from .primitives import *